+++ tests/generate_jwt.js
@@ -1,32 +1,21 @@
-// Make sure to `npm install --save request-promise` or add the dependency to your package.json
-const request = require('request-promise');
+const {JWT} = require('google-auth-library');
+const keys = require('../keys/service-account-key.json');
+console.log(`key:${keys.private_key}/email:${keys.client_email}`)
+async function main() {
+  const client = new JWT(
+    keys.client_email,
+    null,
+    keys.private_key,
+    ['https://www.googleapis.com/auth/cloud-platform'],
+  );
+  const tokenInfo = await client.getTokenInfo(client.credentials.access_token);
+  
+  //const url = `https://dialogflow-demo-api-fd5tyopnsa-an.a.run.app`;
+  const url = `https://dns.googleapis.com/dns/v1/projects/${keys.project_id}`;
+  const tokenInfo = await client.getTokenInfo(client.credentials.access_token);
+  
+  const res = await client.request({url});
+  console.log(res.data);
+}
 
-const receivingServiceURL = 'https://dialogflow-demo-api-fd5tyopnsa-an.a.run.app/';
-
-// Set up metadata server request
-// See https://cloud.google.com/compute/docs/instances/verifying-instance-identity#request_signature
-const metadataServerTokenURL = 'http://metadata/computeMetadata/v1/instance/service-accounts/default/identity?audience=';
-const tokenRequestOptions = {
-    uri: metadataServerTokenURL + receivingServiceURL,
-    headers: {
-        'Metadata-Flavor': 'Google'
-    }
-};
-
-// Fetch the token, then provide the token in the request to the receiving service
-request(tokenRequestOptions)
-  .then((token) => {
-    //console.log(token);
-    return request(receivingServiceURL).auth(null, null, true, token)
-  })
-  .then((response) => {
-    console.log(response);
-    return;
-    res.status(200).send(response);
-  })
-  .catch((error) => {
-    console.log(error);
-    return;
-    res.status(400).send(error);
-  });
-    
+main().catch(console.error);
